Coding time in ms as integers.

+++ read_data.py
@@ -46,7 +46,7 @@ def extract_timeseries(data, tag):
             ] for reading in data[tag]])
 
     timeseries['Time'] = np.asarray(
-        [float(reading[0]) for reading in data[tag]])
+        [int(reading[0]) for reading in data[tag]])
 
     return timeseries
 
@@ -58,6 +58,6 @@ def get_data_ssid(data, ssid):
     data_ssid = list(filter((lambda x: x[2] == ssid), data[tag]))
     timeseries['Data'] = np.asarray([int(reading[-1]) for reading in data_ssid])
 
-    timeseries['Time'] = np.asarray([float(reading[0]) for reading in data_ssid])
+    timeseries['Time'] = np.asarray([int(reading[0]) for reading in data_ssid])
 
     return timeseries
